% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bic.BLMCP.R
\name{plot.bic.BLMCP}
\alias{plot.bic.BLMCP}
\title{Plot coefficients from a "bic.BLMCP" object}
\usage{
\method{plot}{bic.BLMCP}(x, ...)
}
\arguments{
\item{x}{Fitted \code{"bic.BLMCP"} model.}

\item{\dots}{Other graphical parameters to plot.}
}
\value{
A heatmap for estimated coefficients.
}
\description{
Draw a heatmap for estimated coefficients in a fitted
\code{"bic.BLMCP"} object.
}
\references{
Mengyun Wu, Yangguang Zang, Sanguo Zhang, Jian Huang, and Shuangge Ma.
\emph{Accommodating missingness in environmental measurements in gene-environment interaction
analysis. Genetic Epidemiology, 41(6):523-554, 2017.}\cr  Jin Liu, Jian Huang, Yawei Zhang,
Qing Lan, Nathaniel Rothman, Tongzhang Zheng, and Shuangge Ma.
\emph{Identification of gene-environment interactions in cancer studies using penalization.
Genomics, 102(4):189-194, 2013.}
}
\seealso{
\code{predict}, \code{coef} and \code{BLMCP} methods.
}
