% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bic.PTReg.R
\name{plot.bic.PTReg}
\alias{plot.bic.PTReg}
\title{Plot coefficients from a "bic.PTReg" object}
\usage{
\method{plot}{bic.PTReg}(x, ...)
}
\arguments{
\item{x}{Fitted \code{"bic.PTReg"} model.}

\item{\dots}{Other graphical parameters to plot.}
}
\value{
A heat map for estimated coefficients.
}
\description{
Draw a heat map for estimated coefficients in a fitted
\code{"bic.PTReg"} object.
}
\references{
Yaqing Xu, Mengyun Wu, Shuangge Ma, and Syed Ejaz Ahmed.
\emph{Robust gene-environment interaction analysis using penalized trimmed regression. Journal
of Statistical Computation and Simulation, 88(18):3502-3528, 2018.}
}
\seealso{
\code{bic.PTReg}, and \code{predict}, and \code{coef}
methods.
}
