% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Miss.boosting.R
\name{plot.Miss.boosting}
\alias{plot.Miss.boosting}
\title{Plot coefficients from a "Miss.boosting" object}
\usage{
\method{plot}{Miss.boosting}(x, ...)
}
\arguments{
\item{x}{Fitted \code{"Miss.boosting"} model.}

\item{\dots}{Other graphical parameters to plot.}
}
\value{
A heat map for estimated coefficients.
}
\description{
Draw a plot for estimated parameters in a fitted
\code{"Miss.boosting"} object, where a heat map is displaced for selected discrete
environmental (E) effects, genetic (G) effects and their interactions, and a plot for each
of selected continuous E (EC) effect and interactions between EC and G.
}
\references{
Mengyun Wu and Shuangge Ma.
\emph{Robust semiparametric gene-environment interaction analysis using sparse boosting.
Statistics in Medicine, 38(23):4625-4641, 2019.}
}
\seealso{
\code{Miss.boosting}, and \code{predict} methods.
}
