% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.RobSBoosting.R
\name{plot.RobSBoosting}
\alias{plot.RobSBoosting}
\title{Plot coefficients from a "RobSBoosting" object}
\usage{
\method{plot}{RobSBoosting}(x, ...)
}
\arguments{
\item{x}{Fitted \code{"RobSBoosting"} model.}

\item{\dots}{Other graphical parameters to plot.}
}
\value{
Plots for estimated coefficients.
}
\description{
Draw a plot for estimated parameters in a fitted
\code{"RobSBoosting"} object, where a heat map is displaced for selected discrete
environmental (E) effects, genetic (G) effects and their interactions, and a plot for
each of selected continuous E (EC) effect and interactions between EC and G.
}
\references{
Mengyun Wu and Shuangge Ma.
\emph{Robust semiparametric gene-environment interaction analysis using sparse boosting.
Statistics in Medicine, 38(23):4625-4641, 2019.}
}
\seealso{
\code{RobSBoosting}, \code{predict} and \code{coef}
methods.
}
