% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.Miss.boosting.R
\name{predict.Miss.boosting}
\alias{predict.Miss.boosting}
\title{Make predictions from a "Miss.boosting" object}
\usage{
\method{predict}{Miss.boosting}(object, newE, newG, ...)
}
\arguments{
\item{object}{Fitted \code{"Miss.boosting"} object.}

\item{newE}{Matrix of new values for \code{E} at which predictions are to be
made.}

\item{newG}{Matrix of new values for \code{G} at which predictions are to be
made.}

\item{\dots}{Not used. Other arguments to predict.}
}
\value{
The object returned depends on the \dots{} argument which is passed
on to the \code{predict} method for \code{Miss.boosting} objects.
}
\description{
This function makes predictions from a  Miss.boosting model, using the stored
\code{"Miss.boosting"} object.
}
\references{
Mengyun Wu and Shuangge Ma.
\emph{Robust semiparametric gene-environment interaction analysis using sparse boosting.
Statistics in Medicine, 38(23):4625-4641, 2019.}
}
\seealso{
\code{Miss.boosting}, and \code{plot} methods.
}
