% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.RobSBoosting.R
\name{predict.RobSBoosting}
\alias{predict.RobSBoosting}
\title{Make predictions from a "RobSBoosting" object}
\usage{
\method{predict}{RobSBoosting}(object, newE, newG, ...)
}
\arguments{
\item{object}{Fitted \code{"RobSBoosting"} object.}

\item{newE}{Matrix of new values for \code{E} at which predictions are to be
made.}

\item{newG}{Matrix of new values for \code{G} at which predictions are to be
made.}

\item{\dots}{Not used. Other arguments to predict.}
}
\value{
The object returned depends on the \dots{} argument which is passed
on to the \code{predict} method for \code{RobSBoosting} objects.
}
\description{
This function makes predictions from a  RobSBoosting
model, using the stored \code{"RobSBoosting"} object.
}
\references{
Mengyun Wu and Shuangge Ma.
\emph{Robust semiparametric gene-environment interaction analysis using sparse boosting. Statistics in Medicine, 38(23):4625-4641, 2019.}
}
\seealso{
\code{RobSBoosting}, \code{coef}, and \code{plot} methods.
}
