% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Rob_data}
\alias{Rob_data}
\title{A matrix containing the simulated data for \code{RobSBoosting} and \code{Miss.boosting} methods}
\format{
A matrix with 100 rows and 27 variables.
}
\usage{
data(Rob_data)
}
\description{
A matrix containing the simulated genetic (G) effects (the first 20 columns), environmental (E)
effects (column 21 to column 24), continuous response (column 25), logarithm of survival time
(column 26), and censoring indicator (column 27).
}
\examples{
data(Rob_data)
}
\keyword{datasets}
